+++ app/common/model_api.py
@@ -1,0 +1,7 @@
+# 2024/3/7
+# zhangzhong
+# invoke model api
+
+
+def get_avatar_url_by_describe(describe: str) -> str:
+    return "https://www.baidu.com"
